refactor(organize): files in to bin and lib directories
