+++ src/lib/difference.js
