+++ src/lib/rule-finder.js
